[expo-assets] search for key without file extension
