--- packages/expo-asset/src/EmbeddedAssets.web.ts
@@ -1,4 +1,0 @@
-export function getEmbeddedAssetUri(hash: string, type: string | null): string | null {
-  // noop on web
-  return null;
-}
