+++ packages/expo-asset/src/LocalAssets.web.ts
@@ -1,0 +1,4 @@
+export function getLocalAssetUri(hash: string, type: string | null): string | null {
+  // noop on web
+  return null;
+}
